Use snake_case to rename class 'String'
