+++ lib/string.rb
